TP-introduction/producer/requirements.txt
influxdb-client
kafka-python
